updated payload

+++ tools/gulp/tasks/payload.ts
@@ -1,3 +1,4 @@
+import {spawnSync} from 'child_process';
 import { statSync } from 'fs';
 import { sync as glob } from 'glob';
 import { task } from 'gulp';
@@ -8,8 +9,10 @@ import * as firebaseAdmin from 'firebase-admin';
 import {buildConfig} from '../../packages';
 import {openFirebaseDashboardApp} from '../utils/firebase';
 
-import * as util from '../../../scripts/util';
+import {isTravisBuild, isTravisMasterBuild} from '../utils/travis-ci';
 
+
+const request = require('request');
 
 /** Path to the directory where all bundles are living. */
 const bundlesDir = join(buildConfig.outputDir, 'bundles');
@@ -37,17 +40,19 @@ task('payload', ['mosaic:clean-build'], async () => {
     // Print the results to the console, so we can read it from the CI.
     console.log('Payload Results:', JSON.stringify(results, null, 2));
 
-    const firebaseApp = openFirebaseDashboardApp();
-    const database = firebaseApp.database();
+    if (isTravisBuild() && isTravisMasterBuild()) {
+        const firebaseApp = openFirebaseDashboardApp();
+        const database = firebaseApp.database();
 
-    // process.env['CIRCLE_SHA1']!
-    const currentSha = await util.git([`rev-parse --short HEAD`]);
+        const currentSha = process.env['TRAVIS_PULL_REQUEST_SHA']! || process.env['TRAVIS_COMMIT']!;
 
-    await Promise.all([
-        uploadPayloadResults(database, currentSha, results)
-    ]);
+        await Promise.all([
+            uploadPayloadResults(database, currentSha, results),
+            calculatePayloadDiff(database, currentSha, results)
+        ]);
 
-    firebaseApp.delete();
+        firebaseApp.delete();
+    }
 });
 
 /** Uploads the current payload results to the Dashboard database. */
@@ -55,6 +60,71 @@ async function uploadPayloadResults(database: firebaseAdmin.database.Database, c
                                     currentPayload: any) {
 
     await database.ref('payloads').child(currentSha).set(currentPayload);
+}
+
+async function calculatePayloadDiff(database: firebaseAdmin.database.Database, currentSha: string,
+                                    currentPayload: any) {
+
+    const authToken = process.env['FIREBASE_ACCESS_TOKEN']!;
+
+    if (!authToken) {
+        console.error('Cannot calculate Payload diff because there is no "FIREBASE_ACCESS_TOKEN" ' +
+            'environment variable set.');
+
+        return;
+    }
+
+    const previousSha = getCommitFromPreviousPayloadUpload();
+    const previousPayload = await getPayloadResults(database, previousSha);
+
+    if (!previousPayload) {
+        console.warn('There are no previous payload results uploaded. Cannot calculate payload ' +
+            'difference for this job');
+
+        return;
+    }
+
+    console.log(`Comparing payload against payload results from SHA ${previousSha}`);
+
+    // Calculate the payload diffs by subtracting the previous size of the FESM ES2015 bundles.
+    const cdkFullSize = currentPayload.cdk_fesm_2015;
+    const cdkDiff = roundFileSize(cdkFullSize - previousPayload.cdk_fesm_2015);
+
+    const materialFullSize = currentPayload.material_fesm_2015;
+    const materialDiff = roundFileSize(materialFullSize - previousPayload.material_fesm_2015);
+
+    // Set the Github statuses for the packages by sending a HTTP request to the dashboard functions.
+    await Promise.all([
+        updateGithubStatus(currentSha, 'material', materialDiff, materialFullSize, authToken),
+        updateGithubStatus(currentSha, 'cdk', cdkDiff, cdkFullSize, authToken)
+    ]);
+}
+
+async function updateGithubStatus(commitSha: string, packageName: string, packageDiff: number,
+                                  packageFullSize: number, authToken: string) {
+
+    const options = {
+        url: 'https://us-central1-material2-board.cloudfunctions.net/payloadGithubStatus',
+        headers: {
+            'User-Agent': 'Material2/PayloadTask',
+            'auth-token': authToken,
+            'commit-sha': commitSha,
+            'package-name': packageName,
+            'package-full-size': packageFullSize,
+            'package-size-diff': packageDiff
+        }
+    };
+
+    return new Promise((resolve, reject) => {
+        request(options, (err: any, response: any, body: string) => {
+            if (err) {
+                reject(`Dashboard Error ${err.toString()}`);
+            } else {
+                console.info(`Dashboard Response (${response.statusCode}): ${body}`);
+                resolve(response.statusCode);
+            }
+        });
+    });
 }
 
 /** Returns the size of the given library bundle. */
@@ -66,6 +136,44 @@ function getBundleSize(bundleName: string) {
 /** Returns the size of a file in kilobytes. */
 function getFilesize(filePath: string) {
     return statSync(filePath).size / 1000;
+}
+
+/** Gets payload results of the specified commit sha. */
+async function getPayloadResults(database: firebaseAdmin.database.Database, commitSha: string) {
+    const snapshot = await database.ref('payloads').child(commitSha).once('value');
+
+    if (snapshot.exists()) {
+        return snapshot.val();
+    } else {
+        console.error(`There is no payload data uploaded for SHA ${commitSha}`);
+    }
+}
+
+/** Gets the SHA of the commit where the payload was uploaded before this Travis Job started. */
+function getCommitFromPreviousPayloadUpload(): string {
+    if (isTravisMasterBuild()) {
+        const commitRange = process.env['TRAVIS_COMMIT_RANGE']!;
+        // In some situations, Travis will include multiple commits in a single Travis Job. This means
+        // that we can't just resolve the previous commit by using the parent commit of HEAD.
+        // By resolving the amount of commits inside of the current Travis Job, we can figure out
+        // how many commits before HEAD the last Travis Job ran.
+        const commitCount = spawnSync('git', ['rev-list', '--count', commitRange]).stdout
+            .toString().trim();
+
+        // With the amount of commits inside of the current Travis Job, we can query Git to print
+        // the SHA of the commit that ran before this Travis Job was created.
+        return spawnSync('git', ['rev-parse', `HEAD~${commitCount}`]).stdout.toString().trim();
+    } else {
+        // Travis applies the changes of Pull Requests in new branches. This means that resolving
+        // the commit that previously ran on the target branch (mostly "master") can be done
+        // by just loading the SHA of the most recent commit in the target branch.
+        return spawnSync('git', ['rev-parse', process.env['TRAVIS_BRANCH']!]).stdout.toString().trim();
+    }
+}
+
+/** Rounds the specified file size to two decimal places. */
+function roundFileSize(fileSize: number) {
+    return Math.round(fileSize * 100) / 100;
 }
 
 /* tslint:enable:no-magic-numbers */
